saw-core: Move recursor-related code from CtxTerm to SharedTerm.
