+++ saw-core/src/SAWCore/SharedTerm.hs
@@ -780,6 +780,178 @@ getTerm cache termF =
 --------------------------------------------------------------------------------
 -- Recursors
 
+-- | The class of monads that can build terms and substitute into them
+class Monad m => MonadTerm m where
+  mkTermF :: TermF Term -> m Term
+  liftTerm :: DeBruijnIndex -> DeBruijnIndex -> Term -> m Term
+  whnfTerm :: Term -> m Term
+  substTerm :: DeBruijnIndex -> [Term] -> Term -> m Term
+               -- ^ NOTE: the first term in the list is substituted for the most
+               -- recently-bound variable, i.e., deBruijn index 0
+
+-- | Build a 'Term' from a 'FlatTermF' in a 'MonadTerm'
+mkFlatTermF :: MonadTerm m => FlatTermF Term -> m Term
+mkFlatTermF = mkTermF . FTermF
+
+-- | Build a free variable as a 'Term'
+ctxVar :: MonadTerm m => DeBruijnIndex -> m Term
+ctxVar i = mkTermF (LocalVar i)
+
+-- | Build a list of all the free variables as 'Term's
+ctxVars :: MonadTerm m => [(LocalName, tp)] -> m [Term]
+ctxVars = helper
+  where
+    helper :: MonadTerm m => [(LocalName, tp)] -> m [Term]
+    helper [] = pure []
+    helper (_ : ctx) = (:) <$> ctxVar (length ctx) <*> helper ctx
+
+-- | Build two lists of the free variables, split at a specific point
+ctxVars2 :: MonadTerm m => [(LocalName, tp)] ->
+            [(LocalName, tp)] ->
+            m ([Term], [Term])
+ctxVars2 vars1 vars2 =
+  splitAt (length vars1) <$> ctxVars (vars1 ++ vars2)
+
+
+-- | Build a 'Term' for a 'Sort'
+ctxSort :: MonadTerm m => Sort -> m Term
+ctxSort s = mkFlatTermF (Sort s noFlags)
+
+-- | Apply two 'Term's
+ctxApply :: MonadTerm m => m Term -> m Term -> m Term
+ctxApply fm argm =
+  do f <- fm
+     arg <- argm
+     mkTermF (App f arg)
+
+-- | Apply a 'Term' to a list of arguments
+ctxApplyMulti :: MonadTerm m =>
+                 m Term ->
+                 m [Term] ->
+                 m Term
+ctxApplyMulti fm argsm =
+  fm >>= \f -> argsm >>= \args -> helper f args
+  where
+    helper :: MonadTerm m => Term -> [Term] -> m Term
+    helper f [] = return f
+    helper f (arg : args) =
+      do f' <- ctxApply (return f) (return arg)
+         helper f' args
+
+-- | Form a lambda-abstraction as a 'Term'
+ctxLambda1 :: MonadTerm m => LocalName -> Term ->
+              (Term -> m Term) ->
+              m Term
+ctxLambda1 x tp body_f =
+  do var <- ctxVar 0
+     body <- body_f var
+     mkTermF (Lambda x tp body)
+
+-- | Form a multi-arity lambda-abstraction as a 'Term'
+ctxLambda :: MonadTerm m => [(LocalName, Term)] ->
+             ([Term] -> m Term) -> m Term
+ctxLambda [] body_f = body_f []
+ctxLambda ((x, tp) : xs) body_f =
+  ctxLambda1 x tp $ \_ ->
+  ctxLambda xs $ \vars ->
+  do var <- ctxVar (length xs)
+     body_f (var : vars)
+
+-- | Form a pi-abstraction as a 'Term'
+ctxPi1 :: MonadTerm m => LocalName -> Term ->
+          (Term -> m Term) -> m Term
+ctxPi1 x tp body_f =
+  do var <- ctxVar 0
+     body <- body_f var
+     mkTermF (Pi x tp body)
+
+-- | Form a multi-arity pi-abstraction as a 'Term'
+ctxPi :: MonadTerm m => [(LocalName, Term)] ->
+         ([Term] -> m Term) -> m Term
+ctxPi [] body_f = body_f []
+ctxPi ((x, tp) : xs) body_f =
+  ctxPi1 x tp $ \_ ->
+  ctxPi xs $ \vars ->
+  do var <- ctxVar (length xs)
+     body_f (var : vars)
+
+-- | Build an application of a datatype as a 'Term'
+ctxDataTypeM ::
+  forall m.
+  MonadTerm m =>
+  Name ->
+  m [Term] ->
+  m [Term] ->
+  m Term
+ctxDataTypeM d paramsM ixsM =
+  ctxApplyMulti (ctxApplyMulti t paramsM) ixsM
+  where
+    t :: m Term
+    t = mkTermF (Constant d)
+
+-- | Build an application of a constructor as a 'Term'
+ctxCtorAppM ::
+  forall m.
+  MonadTerm m =>
+  Name ->
+  ExtCns Term ->
+  m [Term] ->
+  m [Term] ->
+  m Term
+ctxCtorAppM _d c paramsM argsM =
+  ctxApplyMulti (ctxApplyMulti t paramsM) argsM
+  where
+    t :: m Term
+    t = mkTermF (Constant (Name (ecVarIndex c) (ecName c)))
+
+ctxRecursorAppM :: MonadTerm m =>
+  m Term ->
+  m [Term] ->
+  m Term ->
+  m Term
+ctxRecursorAppM recM ixsM argM =
+  do app <- RecursorApp <$> recM <*> ixsM <*> argM
+     mkFlatTermF app
+
+-- | The class of "in-context" types that support lifting and substitution
+class Monad m => CtxLiftSubst f m where
+  -- | Lift an @f@ into an extended context
+  ctxLift :: DeBruijnIndex -> DeBruijnIndex -> f -> m f
+  -- | Substitute a list of terms into an @f@
+  ctxSubst :: [Term] -> DeBruijnIndex -> f -> m f
+
+instance MonadTerm m => CtxLiftSubst Term m where
+  ctxLift i j t = liftTerm i j t
+  ctxSubst subst i t =
+    -- NOTE: our term lists put the least recently-bound variable first, so we
+    -- have to reverse here to call substTerm, which wants the term for the most
+    -- recently-bound variable first
+    substTerm i (reverse subst) t
+
+instance MonadTerm m => CtxLiftSubst [Term] m where
+  ctxLift i j ts = traverse (ctxLift i j) ts
+  ctxSubst subst i ts = traverse (ctxSubst subst i) ts
+
+instance CtxLiftSubst tp m => CtxLiftSubst [(LocalName, tp)] m where
+  ctxLift _ _ [] = return []
+  ctxLift i j ((x, x_tp) : bs) =
+    (\t -> (:) (x, t)) <$> ctxLift i j x_tp <*>
+    ctxLift (i + 1) j bs
+  ctxSubst _ _ [] = return []
+  ctxSubst subst i ((x, x_tp) : bs) =
+    (\t -> (:) (x, t)) <$> ctxSubst subst i x_tp <*>
+    ctxSubst subst (i + 1) bs
+
+instance MonadTerm m => CtxLiftSubst CtorArg m where
+  ctxLift i j (ConstArg tp) = ConstArg <$> ctxLift i j tp
+  ctxLift i j (RecursiveArg zs ixs) =
+    RecursiveArg <$> ctxLift i j zs <*>
+    ctxLift (i + length zs) j ixs
+  ctxSubst subst i (ConstArg tp) = ConstArg <$> ctxSubst subst i tp
+  ctxSubst subst i (RecursiveArg zs ixs) =
+    RecursiveArg <$> ctxSubst subst i zs <*>
+    ctxSubst subst (i + length zs) ixs
+
 -- | Helper monad for building terms relative to a 'SharedContext'
 newtype ShCtxM a = ShCtxM (ReaderT SharedContext IO a)
                  deriving (Functor, Applicative, Monad)
@@ -812,6 +984,44 @@ allowedElimSort dt s =
     length (dtCtors dt) == 1
   else True
 
+-- | Internal: Convert a 'CtorArg' into the type that it represents,
+-- given a context of the parameters and of the previous arguments.
+ctxCtorArgType :: MonadTerm m => Name ->
+                  [(LocalName, Term)] ->
+                  [(LocalName, Term)] ->
+                  CtorArg ->
+                  m Term
+ctxCtorArgType _ _ _ (ConstArg tp) = return tp
+ctxCtorArgType d params prevs (RecursiveArg zs_ctx ixs) =
+  ctxPi zs_ctx $ \_ ->
+  ctxDataTypeM d ((fst <$> ctxVars2 params prevs) >>= ctxLift 0 (length zs_ctx))
+  (return ixs)
+
+-- | Internal: Convert a bindings list of 'CtorArg's to a binding list
+-- of types.
+ctxCtorArgBindings :: MonadTerm m => Name ->
+                      [(LocalName, Term)] ->
+                      [(LocalName, Term)] ->
+                      [(LocalName, CtorArg)] ->
+                      m [(LocalName, Term)]
+ctxCtorArgBindings _ _ _ [] = return []
+ctxCtorArgBindings d params prevs ((x, arg) : args) =
+  do tp <- ctxCtorArgType d params prevs arg
+     rest <- ctxCtorArgBindings d params (prevs ++ [(x, tp)]) args
+     return ((x, tp) : rest)
+
+-- | Internal: Compute the type of a constructor from the name of its
+-- datatype and its 'CtorArgStruct'
+ctxCtorType :: MonadTerm m => Name -> CtorArgStruct -> m Term
+ctxCtorType d (CtorArgStruct{..}) =
+  (ctxPi ctorParams $ \params ->
+    do bs <-
+         ctxCtorArgBindings d ctorParams
+         [] ctorArgs
+       ctxPi bs $ \_ ->
+         ctxDataTypeM d
+         (ctxLift 0 (length bs) params)
+         (return ctorIndices))
 
 -- | Build a 'Ctor' from a 'CtorArgStruct' and a list of the other constructor
 -- names of the 'DataType'. Note that we cannot look up the latter information,
@@ -871,6 +1081,288 @@ scBuildCtor sc d c arg_struct =
       , ctorIotaTemplate  = iota_red
       , ctorIotaReduction = iota_fun
       }
+
+-- | Build the type of the @p_ret@ function, also known as the "motive"
+-- function, of a recursor on datatype @d@. This type has the form
+--
+-- > (i1::ix1) -> .. -> (im::ixm) -> d p1 .. pn i1 .. im -> s
+--
+-- where the @pi@ are free variables for the parameters of @d@, the @ixj@
+-- are the indices of @d@, and @s@ is any sort supplied as an argument.
+ctxPRetTp :: MonadTerm m => Name ->
+             [(LocalName, Term)] ->
+             [(LocalName, Term)] -> Sort ->
+             m Term
+ctxPRetTp d params ixs s =
+  ctxPi ixs $ \ix_vars ->
+  do param_vars <- ctxVars params
+     dt <- ctxDataTypeM d (ctxLift 0 (length ixs) param_vars)
+       (return ix_vars)
+     ctxPi1 "_" dt $ \_ -> ctxSort s
+
+-- | Take a list of terms and match them up with a sequence of bindings,
+-- returning a structured '[Term]' list.
+ctxTermsForBindings :: [(LocalName, tp)] -> [Term] -> Maybe [Term]
+ctxTermsForBindings bs ts
+  | length bs == length ts = Just ts
+  | otherwise = Nothing
+
+-- | Like 'ctxPRetTp', but also take in a list of parameters and substitute them
+-- for the parameter variables returned by that function
+mkPRetTp :: MonadTerm m =>
+  Name ->
+  [(LocalName, Term)] ->
+  [(LocalName, Term)] ->
+  [Term] ->
+  Sort ->
+  m Term
+mkPRetTp d p_ctx ix_ctx untyped_params s =
+  case ctxTermsForBindings p_ctx untyped_params of
+    Just params ->
+      do p_ret <- ctxPRetTp d p_ctx ix_ctx s
+         ctxSubst params 0 p_ret
+    Nothing ->
+      error "mkPRetTp: incorrect number of parameters"
+
+
+-- | Compute the type of an eliminator function for a constructor from the name
+-- of its datatype, its name, and its 'CtorArgStruct'. This type has, as free
+-- variables, both the parameters of the datatype and a "motive" function from
+-- indices of the datatype to a return type. It is of the form
+--
+-- > (x1::arg1) -> maybe (rec1::rec_tp1) -> .. ->
+-- > (xn::argn) -> maybe (recn::rec_tpn) ->
+-- >   p_ret ix_1 .. ix_k (ctor params x1 .. xn)
+--
+-- where the ixs are the type indices of the return type for the constructor,
+-- the (xi::argi) are the arguments of the constructor, and the @maybe@s
+-- indicate additional arguments that are present only for arguments of
+-- recursive type, that is, where @argi@ has the form
+--
+-- > (z1::Z1) -> .. -> (zm::Zm) -> d params t1 .. tk
+--
+-- In this case, @rec_tpi@ has the form
+--
+-- > (z1::Z1) -> .. -> (zm::Zm) -> p_ret t1 .. tk (f z1 .. zm)
+--
+-- Note that the output type cannot be expressed in the type of this function,
+-- since it depends on fields of the 'CtorArgStruct', so, instead, the result is
+-- just casted to whatever type the caller specifies.
+ctxCtorElimType :: MonadTerm m =>
+  Name ->
+  ExtCns Term ->
+  CtorArgStruct ->
+  m Term
+ctxCtorElimType d_top c
+  (CtorArgStruct{..}) =
+  (do let params = ctorParams
+      -- NOTE: we use propSort for the type of p_ret just as arbitrary sort, but
+      -- it doesn't matter because p_ret_tp is only actually used to form
+      -- contexts, and is never actually used directly in the output
+      p_ret_tp <- ctxPRetTp d_top params dataTypeIndices propSort
+
+      -- Lift the argument and return indices into the context of p_ret
+      args <- ctxLift 0 1 ctorArgs
+      ixs <-
+        ctxLift (length ctorArgs) 1
+        ctorIndices
+      -- Form the context (params ::> p_ret)
+      let pret = ("_", p_ret_tp)
+      -- Call the helper and cast the result to (Typ ret)
+      helper d_top params pret [] args ixs
+  ) where
+
+  -- Iterate through the argument types of the constructor, building up a
+  -- function from those arguments to the result type of the p_ret function.
+  -- Note that, technically, this function also takes in recursive calls, so has
+  -- a slightly richer type, but we are not going to try to compute this richer
+  -- type in Haskell land.
+  helper :: MonadTerm m =>
+    Name ->
+    [(LocalName, Term)] ->
+    (LocalName, Term) ->
+    [(LocalName, Term)] ->
+    [(LocalName, CtorArg)] ->
+    [Term] ->
+    m Term
+  helper d params pret prevs [] ret_ixs =
+    -- If we are finished with our arguments, construct the final result type
+    -- (p_ret ret_ixs (c params prevs))
+    do (vars, prev_vars) <- ctxVars2 (params ++ [pret]) prevs
+       -- note: length vars == length (params ++ [pret])
+       let param_terms = init vars
+       let p_ret = last vars
+       ctxApply (ctxApplyMulti (return p_ret) (return ret_ixs)) $
+         ctxCtorAppM d c (return param_terms) (return prev_vars)
+  helper d params pret prevs ((str, ConstArg tp) : args) ixs =
+    -- For a constant argument type, just abstract it and continue
+    (ctxPi [(str, tp)] $ \_ ->
+      helper d params pret (prevs ++ [(str, tp)]) args ixs)
+  helper d params pret
+    prevs ((str, RecursiveArg zs ts) : args) ixs =
+    -- For a recursive argument type of the form
+    --
+    -- (z1::Z1) -> .. -> (zm::Zm) -> d params t1 .. tk
+    --
+    -- form the type abstraction
+    --
+    -- (arg:: (z1::Z1) -> .. -> (zm::Zm) -> d params t1 .. tk) ->
+    -- (ih :: (z1::Z1) -> .. -> (zm::Zm) -> p_ret t1 .. tk (arg z1 .. zm)) ->
+    -- rest
+    --
+    -- where rest is the result of a recursive call
+    do
+      -- Build terms for the params and p_ret variables
+      vars <- fst <$> ctxVars2 (params ++ [pret]) prevs
+      -- note: length vars == length (params ++ [pret])
+      let param_vars = init vars
+      let p_ret = last vars
+      -- Build the type of the argument arg
+      arg_tp <- ctxPi zs (\_ -> ctxDataTypeM d
+                                (ctxLift 0 (length zs) param_vars)
+                                (return ts))
+      -- Lift zs and ts into the context of arg
+      zs' <- ctxLift 0 1 zs
+      ts' <- ctxLift (length zs) 1 ts
+      -- Build the pi-abstraction for arg
+      ctxPi1 str arg_tp $ \arg ->
+        do rest <-
+             helper d params pret (prevs ++ [(str, arg_tp)]) args ixs
+           -- Build the type of ih, in the context of arg
+           ih_tp <- ctxPi zs' $ \z_vars ->
+             ctxApply
+             (ctxApplyMulti
+              (ctxLift 0 (length zs' + 1) p_ret) (return ts'))
+             (ctxApplyMulti (ctxLift 0 (length zs') arg) (return z_vars))
+           -- Finally, build the pi-abstraction for ih around the rest
+           --
+           -- NOTE: we cast away the IH argument, because that is a type that is
+           -- computed from the argument structure, and we cannot (well, we
+           -- could, but it would be much more work to) express that computation
+           -- in the Haskell type system
+           (ctxPi1 "_" ih_tp $ \_ ->
+               ctxLift 0 1 rest)
+
+-- | Build a function that substitutes parameters and a @p_ret@ return type
+-- function into the type of an eliminator, as returned by 'ctxCtorElimType',
+-- for the given constructor. We return the substitution function in the monad
+-- so that we only call 'ctxCtorElimType' once but can call the function many
+-- times, in order to amortize the overhead of 'ctxCtorElimType'.
+mkCtorElimTypeFun :: MonadTerm m =>
+  Name {- ^ data type -} ->
+  ExtCns Term {- ^ constructor type -} ->
+  CtorArgStruct ->
+  m ([Term] -> Term -> m Term)
+mkCtorElimTypeFun d c argStruct@(CtorArgStruct {..}) =
+  do ctxElimType <- ctxCtorElimType d c argStruct
+     return $ \params p_ret ->
+         whnfTerm =<<
+         case ctxTermsForBindings ctorParams params of
+           Nothing -> error "ctorElimTypeFun: wrong number of parameters!"
+           Just paramsCtx ->
+             ctxSubst
+             (paramsCtx ++ [p_ret])
+             0 ctxElimType
+
+-- | Zip two lists of equal length, but return 'Nothing' if the
+-- lengths are different.
+zipSameLength :: [a] -> [b] -> Maybe [(a, b)]
+zipSameLength xs ys
+  | length xs == length ys = Just (zip xs ys)
+  | otherwise = Nothing
+
+-- | Reduce an application of a recursor to a particular constructor.
+-- This is known in the Coq literature as an iota reduction. More specifically,
+-- the call
+--
+-- > ctxReduceRecursor rec f_c [x1, .., xk]
+--
+-- reduces the term @(RecursorApp rec ixs (CtorApp c ps xs))@ to
+--
+-- > f_c x1 (maybe rec_tm_1) .. xk (maybe rec_tm_k)
+--
+-- where @f_c@ is the eliminator function associated to the constructor @c@ by the
+-- recursor value @rec@.  Here @maybe rec_tm_i@ indicates an optional recursive call
+-- of the recursor on one of the @xi@. These recursive calls only exist for those
+-- arguments @xi@ that are recursive arguments, i.e., that are specified with
+-- 'RecursiveArg', and are omitted for non-recursive arguments specified by 'ConstArg'.
+--
+-- Specifically, for a @'RecursiveArg' zs ixs@ argument @xi@, which has type
+-- @\(z1::Z1) -> .. -> d p1 .. pn ix1 .. ixp@, we build the recursive call
+--
+-- > \(z1::[xs/args]Z1) -> .. ->
+-- >   RecursorApp rec [xs/args]ixs (xi z1 ... zn)
+--
+-- where @[xs/args]@ substitutes the concrete values for the earlier
+-- constructor arguments @xs@ for the remaining free variables.
+
+ctxReduceRecursor :: forall m.
+  MonadTerm m =>
+  Term {- ^ abstracted recursor -} ->
+  Term {- ^ constructor elimnator function -} ->
+  [Term] {- ^ constructor arguments -} ->
+  CtorArgStruct {- ^ constructor formal argument descriptor -} ->
+  m Term
+ctxReduceRecursor rec elimf c_args CtorArgStruct{..} =
+  case zipSameLength c_args ctorArgs of
+     Just argsCtx_ctorArgs ->
+       ctxReduceRecursor_ rec elimf argsCtx_ctorArgs
+     Nothing ->
+       error "ctxReduceRecursorRaw: wrong number of constructor arguments!"
+
+
+-- | This operation does the real work of building the
+--   iota reduction for @ctxReduceRecursor@.  We assume
+--   the input terms we are given live in an ambient
+--   context @amb@.
+ctxReduceRecursor_ :: forall m.
+  MonadTerm m =>
+  Term     {- ^ recursor value eliminatiting data type d -}->
+  Term     {- ^ eliminator function for the constructor -} ->
+  [(Term, (LocalName, CtorArg))] {- ^ constructor actual arguments plus argument descriptions -} ->
+  m Term
+ctxReduceRecursor_ rec fi args0_argCtx =
+  do args <- mk_args [] args0_argCtx
+     whnfTerm =<< foldM (\f arg -> mkTermF $ App f arg) fi args
+
+ where
+    mk_args :: [Term] ->  -- already processed parameters/arguments
+               [(Term, (LocalName, CtorArg))] ->
+                 -- remaining actual arguments to process, with
+                 -- telescope for typing the actual arguments
+               m [Term]
+    -- no more arguments to process
+    mk_args _ [] = return []
+
+    -- process an argument that is not a recursive call
+    mk_args pre_xs ((x, (_, ConstArg _)) : xs_args) =
+      do tl <- mk_args (pre_xs ++ [x]) xs_args
+         pure (x : tl)
+
+    -- process an argument that is a recursive call
+    mk_args pre_xs ((x, (_, RecursiveArg zs ixs)) : xs_args) =
+      do zs'  <- ctxSubst pre_xs 0 zs
+         ixs' <- ctxSubst pre_xs (length zs) ixs
+         recx <- mk_rec_arg zs' ixs' x
+         tl   <- mk_args (pre_xs ++ [x]) xs_args
+         pure (x : recx : tl)
+
+    -- Build an individual recursive call, given the parameters, the bindings
+    -- for the RecursiveArg, and the argument we are going to recurse on
+    mk_rec_arg ::
+      [(LocalName, Term)] ->                -- telescope describing the zs
+      [Term] ->                        -- actual values for the indices, shifted under zs
+      Term ->                         -- actual value in recursive position
+      m Term
+    mk_rec_arg zs_ctx ixs x =
+      -- eta expand over the zs and apply the RecursorApp form
+      ctxLambda zs_ctx (\zs ->
+        ctxRecursorAppM
+          (ctxLift 0 (length zs_ctx) rec)
+          (return ixs)
+          (ctxApplyMulti
+            (ctxLift 0 (length zs_ctx) x)
+            (return zs)))
 
 -- | Given a datatype @d@, parameters @p1,..,pn@ for @d@, and a "motive"
 -- function @p_ret@ of type
